Split integration_calculations.cc into proper src/ modules
The integrator and baseline both source-included
integration_calculations.cc, a catch-all compiled into each translation
unit.

Split it into two logically-named modules following the existing
extent.cc pattern (declaration in include/integrator/, definition in
src/integrator/), compiled and linked rather than #included:

- lp_correction.{hpp,cc}    - lorentz/polarization correction,
  LPCorrection
- coordinate_system.{hpp,cc} - CoordinateSystem (Kabsch frame)

Both headers carry full Doxygen documentation. integrator.cc and
baseline/integrator/integrator.cc now include the headers; both
CMakeLists compile the new sources alongside extent.cc.
integration_calculations.cc is removed.
