+++ baseline/integrator/CMakeLists.txt
@@ -6,6 +6,8 @@ set(CMAKE_CXX_STANDARD 20)
 add_executable(baseline_integrator
     integrator.cc
     ${CMAKE_SOURCE_DIR}/src/integrator/extent.cc
+    ${CMAKE_SOURCE_DIR}/src/integrator/lp_correction.cc
+    ${CMAKE_SOURCE_DIR}/src/integrator/coordinate_system.cc
 )
 
 target_include_directories(baseline_integrator PRIVATE ${CMAKE_SOURCE_DIR}/baseline/predictor)
