+++ baseline/integrator/integrator.cc
@@ -25,8 +25,9 @@
 #include "common.hpp"
 #include "ffs_logger.hpp"
 #include "h5read.h"
-#include "integration_calculations.cc"
+#include "integrator/coordinate_system.hpp"
 #include "integrator/extent.hpp"
+#include "integrator/lp_correction.hpp"
 #include "kabsch.cc"
 #include "math/math_utils.cuh"
 #include "predict.cc"
