+++ integrator/CMakeLists.txt
@@ -9,6 +9,8 @@ add_executable(integrator
     integrator.cu
     kabsch.cu
     ${CMAKE_SOURCE_DIR}/src/integrator/extent.cc
+    ${CMAKE_SOURCE_DIR}/src/integrator/lp_correction.cc
+    ${CMAKE_SOURCE_DIR}/src/integrator/coordinate_system.cc
     ../spotfinder/shmread.cc
     ../spotfinder/cbfread.cc
 )
