+++ integrator/integrator.cc
@@ -37,9 +37,10 @@
 #include "cuda_common.hpp"
 #include "ffs_logger.hpp"
 #include "h5read.h"
-#include "integration_calculations.cc"
 #include "integrator.cuh"
+#include "integrator/coordinate_system.hpp"
 #include "integrator/extent.hpp"
+#include "integrator/lp_correction.hpp"
 #include "kabsch.cuh"
 #include "math/device_precision.cuh"
 #include "math/math_utils.cuh"
